+++ lib/holder/MyPage/user/components/user_my_page_pic.dart
@@ -1,0 +1,39 @@
+
+
+
+
+import 'package:flutter/cupertino.dart';
+
+import '../../../../theme.dart';
+
+class user_my_page_pic extends StatelessWidget {
+  const user_my_page_pic({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        SizedBox(
+          width: 65,
+          height: 65,
+          child: ClipRRect(
+            borderRadius: BorderRadius.circular(32.5),
+            child: Image.network(
+              'https://picsum.photos/200/100', // :TODO 04 사진수정
+              fit: BoxFit.cover,
+            ),
+          ),
+        ),
+        SizedBox(width: 16),
+        Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text('회원 이름', style: textTheme().displayMedium), // :TODO 04수정
+          ],
+        ),
+      ],
+    );
+  }
+}
