+++ lib/holder/MyPage/user/components/user_tab_bar_grid.dart
@@ -1,0 +1,15 @@
+import 'package:final_project_team02/holder/MyPage/user/components/user_apply_button.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
+
+class UserTabGird extends StatelessWidget {
+  const UserTabGird({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Align(child: Text("크리에이터가 되어 자신의 멋진 코디를 뽐내 보세요.!!"));
+  }
+}
