+++ lib/holder/MyPage/user/user_my_page.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 
 import '../../../components/custom_app_bar.dart';
+import '../creator/components/creator_my_page_header.dart';
 import 'components/user_my_page_body.dart';
+import 'components/user_my_page_bottom.dart';
 import 'components/user_my_page_header.dart';
 
 class UserMyPage extends StatelessWidget {
@@ -10,17 +12,8 @@ class UserMyPage extends StatelessWidget {
     return Scaffold(
       appBar: CustomAppBar(miconRight: Icons.home, miconLeft: Icons.settings),
       backgroundColor: Colors.white,
-      body: NestedScrollView(
-        headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled) {
-          return [
-            SliverToBoxAdapter(
-              //프로필 설정 버튼까지
-              child: MyPageheader(),
-            ),
-          ];
-        },
-        body: MyPageBody(),
-      ),
+      body: UserMyPageBody(),
     );
   }
 }
+
